requirements/base.txt
celery[redis]==4.2.1
channels==2.1.3
channels_redis==2.3.0
django[argon2]==2.1
django-cors-headers==2.4.0
django-environ==0.4.5
django-filter==2.0.0
django-redis==4.9.0
djangorestframework==3.8.2
djangorestframework_simplejwt==3.2.3
drf-yasg[validation]==1.10.0
gunicorn==19.9.0
Markdown==2.6.11
openapi==1.1.0
pika==0.12.0
psycopg2==2.7.5
Pygments==2.2.0
uvicorn==0.3.2
